refactor: rename class
